review tree layout for tx/certificates
