+++ tx/certificates/errors.go
@@ -1,0 +1,10 @@
+package certificates
+
+import (
+	"errors"
+)
+
+var (
+	UnknownCertKind = errors.New("Unkown certificate kind")
+	WrongCertKind   = errors.New("Wrong certificate kind")
+)
